Move node customisers and node sources into common
